third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-expected.txt

CONSOLE ERROR: line 23: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 44: Uncaught (in promise) Error: err2
CONSOLE ERROR: line 62: Uncaught (in promise) NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
CONSOLE ERROR: Uncaught (in promise) 42
CONSOLE ERROR: line 73: Uncaught (in promise) 1e+100
CONSOLE ERROR: line 78: Uncaught (in promise) foo
CONSOLE ERROR: line 83: Uncaught (in promise) #<Object>
CONSOLE ERROR: line 88: Uncaught (in promise) undefined
Tests that uncaught promise rejections are logged into console.


    console-uncaught-promise.html:23 Uncaught (in promise) Error: err1 {stack: (...), message: "err1"}
        message: "err1"
        stack: (...)
        get stack: function () { [native code] }
        set stack: function () { [native code] }
        __proto__: d
        console-uncaught-promise.html:23 promiseTest1
        console-uncaught-promise.html:16 timeout

    console-uncaught-promise.html:44 Uncaught (in promise) Error: err2 {stack: (...), message: "err2"}
        message: "err2"
        stack: (...)
        get stack: function () { [native code] }
        set stack: function () { [native code] }
        __proto__: d
        console-uncaught-promise.html:44 promiseTest2
        console-uncaught-promise.html:16 timeout

    console-uncaught-promise.html:62 Uncaught (in promise) DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node. {...<removed>...}
        code: 8
        message: "Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node."
        name: "NotFoundError"
        stack: (...)
        __proto__: DOMException
        console-uncaught-promise.html:62 throwDOMException
        console-uncaught-promise.html:54 catcher
console-uncaught-promise.html:1 Uncaught (in promise) 42

    console-uncaught-promise.html:73 Uncaught (in promise) 1e+100
        console-uncaught-promise.html:73 promiseTest5
        console-uncaught-promise.html:16 timeout

    console-uncaught-promise.html:78 Uncaught (in promise) foo
        console-uncaught-promise.html:78 promiseTest6
        console-uncaught-promise.html:16 timeout

    console-uncaught-promise.html:83 Uncaught (in promise) Object {foo: 42}
        console-uncaught-promise.html:83 promiseTest7
        console-uncaught-promise.html:16 timeout

    console-uncaught-promise.html:88 Uncaught (in promise) undefined
        console-uncaught-promise.html:88 promiseTest8
        console-uncaught-promise.html:16 timeout
